requirements.txt
# Core dependencies
pyyaml>=6.0
pillow>=10.0.0
numpy>=1.24.0

# Audio processing
gtts>=2.3.0
pydub>=0.25.1

# Video processing
moviepy>=1.0.3

# CLI
click>=8.1.0

# Optional AI dependencies (fallback if unavailable)
# transformers>=4.30.0
# diffusers>=0.20.0
# torch>=2.0.0
